Migrate to plugin/index ts files.
